@@ -89,23 +86,23 @@
     "msw": "0.49.3",
     "omit-empty-es": "1.2.0",
     "prettier": "2.8.8",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-intl": "^6.4.7",
+   "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-intl": "^7.1.4",
     "react-redux": "7.2.9",
     "react-router-dom": "5.3.4",
-    "recharts": "^2.15.3",
     "redux": "4.2.1",
+    "recharts": "^2.15.3",
     "typescript": "5.0.4"
   },
   "resolutions": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@types/eslint": "<9",
-    "@types/react": "<18",
-    "@types/react-dom": "<18",
     "@types/react-router": "<6",
     "@types/react-router-dom": "<6",
+    "@types/eslint": "<9",
+    "@types/react": ">18",
+    "@types/react-dom": ">18",
     "headers-polyfill": "3.2.5",
     "nwsapi": "2.2.7"
   },
