@@ -1,6 +1,7 @@
+import { createRoot } from "react-dom/client";
 /// <reference types="@commercetools-frontend/application-config/client" />
 
-import ReactDOM from 'react-dom';
 import EntryPoint from './components/entry-point';
 
-ReactDOM.render(<EntryPoint />, document.getElementById('app'));
+const root = createRoot(document.getElementById('app') as HTMLElement);
+root.render(<EntryPoint />);
